+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased] (date goes here)
 
+### Fixed
+
+- command line tab completion (was only working  when logging)
+
 ## [3.3.0] (2018-10-21)
 
 ### Changed